Remove Base class and move its functionality to Group
