+++ src/python/module/z5py/dataset.py
@@ -7,7 +7,8 @@ import numpy as np
 from ._z5py import open_dataset
 from ._z5py import write_subarray, write_scalar, read_subarray, convert_array_to_format
 from .attribute_manager import AttributeManager
-from .shape_utils import slice_to_begin_shape, int_to_begin_shape, rectify_shape, get_default_chunks
+from .shape_utils import slice_to_begin_shape, int_to_begin_shape, rectify_shape
+from .shape_utils import get_default_chunks, is_group
 
 
 class Dataset(object):
@@ -172,6 +173,10 @@ class Dataset(object):
                         chunks, n_threads,
                         is_zarr, mode, **kwargs):
         if os.path.exists(path):
+
+            if is_group(path, is_zarr):
+                raise TypeError("Incompatible object (Group) already exists")
+
             ds = cls(path, open_dataset(path, mode), n_threads)
             if shape != ds.shape:
                 raise TypeError("Shapes do not match (existing (%s) vs new (%s))" % (', '.join(map(str, ds.shape)),
@@ -184,6 +189,7 @@ class Dataset(object):
                 if np.dtype(dtype) != ds.dtype:
                     raise TypeError("Datatypes do not match (existing %s vs new %s)" % str(ds.dtype), str(dtype))
             return ds
+
         else:
             return cls.create_dataset(path, shape, dtype,
                                       chunks=chunks,
