+++ src/python/module/z5py/shape_utils.py
@@ -1,3 +1,7 @@
+import os
+import json
+
+
 def slice_to_begin_shape(s, size):
     """For a single dimension with a given size, turn a slice object into a (start_idx, length)
      pair. Returns (None, 0) if slice is invalid."""
@@ -98,6 +102,22 @@ def rectify_shape(arr, required_shape):
 def get_default_chunks(shape):
     # the default size is 64**3
     default_size = 262144
-    ndim = len(shape)
-    default_dim = default_size ** (1. / ndim)
+    default_dim = int(round(default_size ** (1. / len(shape))))
     return tuple(min(default_dim, sh) for sh in shape)
+
+
+def is_group(path, is_zarr):
+    if is_zarr:
+        return os.path.exists(os.path.join(path, '.zgroup'))
+    else:
+        meta_path = os.path.join(path, 'attributes.json')
+        if not os.path.exists(meta_path):
+            return True
+        with open(meta_path, 'r') as f:
+            # attributes for n5 file can be empty which cannot be parsed by json
+            try:
+                attributes = json.load(f)
+            except ValueError:
+                attributes = {}
+        # The dimensions key is only present in a dataset
+        return 'dimensions' not in attributes
